Revert "[Android] Update fastdeploy Android app ui module"
