+++ java/android/app/src/main/AndroidManifest.xml
@@ -15,14 +15,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ocr.OcrMainActivity">
+        <activity android:name="com.baidu.paddle.fastdeploy.app.examples.ocr.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
         <activity
-            android:name=".ocr.OcrSettingsActivity"
+            android:name="com.baidu.paddle.fastdeploy.app.examples.ocr.SettingsActivity"
             android:label="Settings">
         </activity>
     </application>
