+++ java/android/app/src/main/java/com/baidu/paddle/fastdeploy/app/examples/detection/MainActivity.java
@@ -1,0 +1,255 @@
+package com.baidu.paddle.fastdeploy.app.examples.detection;
+
+
+import android.Manifest;
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.graphics.Bitmap;
+import android.os.Bundle;
+import android.preference.PreferenceManager;
+import android.support.annotation.NonNull;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
+import android.view.View;
+import android.view.Window;
+import android.view.WindowManager;
+import android.widget.ImageButton;
+import android.widget.ImageView;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.baidu.paddle.fastdeploy.RuntimeOption;
+import com.baidu.paddle.fastdeploy.app.examples.R;
+import com.baidu.paddle.fastdeploy.app.ui.CameraSurfaceView;
+import com.baidu.paddle.fastdeploy.app.ui.Utils;
+import com.baidu.paddle.fastdeploy.vision.DetectionResult;
+import com.baidu.paddle.fastdeploy.vision.detection.PicoDet;
+
+import java.io.File;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class MainActivity extends Activity implements View.OnClickListener, CameraSurfaceView.OnTextureChangedListener {
+    private static final String TAG = MainActivity.class.getSimpleName();
+
+    CameraSurfaceView svPreview;
+    TextView tvStatus;
+    ImageButton btnSwitch;
+    ImageButton btnShutter;
+    ImageButton btnSettings;
+    ImageView realtimeToggleButton;
+    boolean isRealtimeStatusRunning = false;
+    ImageView backInPreview;
+
+    String savedImagePath = "result.jpg";
+    int lastFrameIndex = 0;
+    long lastFrameTime;
+
+    // Call 'init' and 'release' manually later
+    PicoDet predictor = new PicoDet();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        // Fullscreen
+        requestWindowFeature(Window.FEATURE_NO_TITLE);
+        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
+
+        setContentView(R.layout.activity_main);
+
+        // Clear all setting items to avoid app crashing due to the incorrect settings
+        initSettings();
+
+        // Init the camera preview and UI components
+        initView();
+
+        // Check and request CAMERA and WRITE_EXTERNAL_STORAGE permissions
+        if (!checkAllPermissions()) {
+            requestAllPermissions();
+        }
+    }
+
+    @SuppressLint("NonConstantResourceId")
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.btn_switch:
+                svPreview.switchCamera();
+                break;
+            case R.id.btn_shutter:
+                @SuppressLint("SimpleDateFormat")
+                SimpleDateFormat date = new SimpleDateFormat("yyyy_MM_dd_HH_mm_ss");
+                synchronized (this) {
+                    savedImagePath = Utils.getDCIMDirectory() + File.separator + date.format(new Date()).toString() + ".png";
+                }
+                Toast.makeText(MainActivity.this, "Save snapshot to " + savedImagePath, Toast.LENGTH_SHORT).show();
+                break;
+            case R.id.btn_settings:
+                startActivity(new Intent(MainActivity.this, SettingsActivity.class));
+                break;
+            case R.id.realtime_toggle_btn:
+                toggleRealtimeStyle();
+                break;
+            case R.id.back_in_preview:
+                finish();
+                break;
+        }
+    }
+
+    private void toggleRealtimeStyle() {
+        if (isRealtimeStatusRunning) {
+            isRealtimeStatusRunning = false;
+            realtimeToggleButton.setImageResource(R.drawable.realtime_stop_btn);
+            svPreview.setOnTextureChangedListener(this);
+            tvStatus.setVisibility(View.VISIBLE);
+        } else {
+            isRealtimeStatusRunning = true;
+            realtimeToggleButton.setImageResource(R.drawable.realtime_start_btn);
+            tvStatus.setVisibility(View.GONE);
+            svPreview.setOnTextureChangedListener(new CameraSurfaceView.OnTextureChangedListener() {
+                @Override
+                public boolean onTextureChanged(Bitmap ARGB8888ImageBitmap) {
+                    return false;
+                }
+            });
+        }
+    }
+
+    @Override
+    public boolean onTextureChanged(Bitmap ARGB8888ImageBitmap) {
+        String savedImagePath = "";
+        synchronized (this) {
+            savedImagePath = MainActivity.this.savedImagePath;
+        }
+        boolean modified = false;
+        DetectionResult result = predictor.predict(
+                ARGB8888ImageBitmap, savedImagePath, SettingsActivity.scoreThreshold);
+        modified = result.initialized();
+        if (!savedImagePath.isEmpty()) {
+            synchronized (this) {
+                MainActivity.this.savedImagePath = "result.jpg";
+            }
+        }
+        lastFrameIndex++;
+        if (lastFrameIndex >= 30) {
+            final int fps = (int) (lastFrameIndex * 1e9 / (System.nanoTime() - lastFrameTime));
+            runOnUiThread(new Runnable() {
+                @SuppressLint("SetTextI18n")
+                public void run() {
+                    tvStatus.setText(Integer.toString(fps) + "fps");
+                }
+            });
+            lastFrameIndex = 0;
+            lastFrameTime = System.nanoTime();
+        }
+        return modified;
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        // Reload settings and re-initialize the predictor
+        checkAndUpdateSettings();
+        // Open camera until the permissions have been granted
+        if (!checkAllPermissions()) {
+            svPreview.disableCamera();
+        }
+        svPreview.onResume();
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        svPreview.onPause();
+    }
+
+    @Override
+    protected void onDestroy() {
+        if (predictor != null) {
+            predictor.release();
+        }
+        super.onDestroy();
+    }
+
+    public void initView() {
+        svPreview = (CameraSurfaceView) findViewById(R.id.sv_preview);
+        svPreview.setOnTextureChangedListener(this);
+        tvStatus = (TextView) findViewById(R.id.tv_status);
+        btnSwitch = (ImageButton) findViewById(R.id.btn_switch);
+        btnSwitch.setOnClickListener(this);
+        btnShutter = (ImageButton) findViewById(R.id.btn_shutter);
+        btnShutter.setOnClickListener(this);
+        btnSettings = (ImageButton) findViewById(R.id.btn_settings);
+        btnSettings.setOnClickListener(this);
+        realtimeToggleButton = findViewById(R.id.realtime_toggle_btn);
+        realtimeToggleButton.setOnClickListener(this);
+        backInPreview = findViewById(R.id.back_in_preview);
+        backInPreview.setOnClickListener(this);
+    }
+
+    @SuppressLint("ApplySharedPref")
+    public void initSettings() {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.clear();
+        editor.commit();
+        SettingsActivity.resetSettings();
+    }
+
+    public void checkAndUpdateSettings() {
+        if (SettingsActivity.checkAndUpdateSettings(this)) {
+            String realModelDir = getCacheDir() + "/" + SettingsActivity.modelDir;
+            Utils.copyDirectoryFromAssets(this, SettingsActivity.modelDir, realModelDir);
+            String realLabelPath = getCacheDir() + "/" + SettingsActivity.labelPath;
+            Utils.copyFileFromAssets(this, SettingsActivity.labelPath, realLabelPath);
+
+            String modelFile = realModelDir + "/" + "model.pdmodel";
+            String paramsFile = realModelDir + "/" + "model.pdiparams";
+            String configFile = realModelDir + "/" + "infer_cfg.yml";
+            String labelFile = realLabelPath;
+            RuntimeOption option = new RuntimeOption();
+            option.setCpuThreadNum(SettingsActivity.cpuThreadNum);
+            option.setLitePowerMode(SettingsActivity.cpuPowerMode);
+            option.enableRecordTimeOfRuntime();
+            if (Boolean.parseBoolean(SettingsActivity.enableLiteFp16)) {
+                option.enableLiteFp16();
+            }
+            predictor.init(modelFile, paramsFile, configFile, labelFile, option);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (grantResults[0] != PackageManager.PERMISSION_GRANTED || grantResults[1] != PackageManager.PERMISSION_GRANTED) {
+            new AlertDialog.Builder(MainActivity.this)
+                    .setTitle("Permission denied")
+                    .setMessage("Click to force quit the app, then open Settings->Apps & notifications->Target " +
+                            "App->Permissions to grant all of the permissions.")
+                    .setCancelable(false)
+                    .setPositiveButton("Exit", new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            MainActivity.this.finish();
+                        }
+                    }).show();
+        }
+    }
+
+    private void requestAllPermissions() {
+        ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                Manifest.permission.CAMERA}, 0);
+    }
+
+    private boolean checkAllPermissions() {
+        return ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
+                && ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED;
+    }
+}
