+++ java/android/app/src/main/java/com/baidu/paddle/fastdeploy/app/examples/ocr/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.baidu.paddle.fastdeploy.app.examples.detection;
+package com.baidu.paddle.fastdeploy.app.examples.ocr;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -10,15 +10,15 @@ import android.preference.PreferenceManager;
 import android.support.v7.app.ActionBar;
 
 import com.baidu.paddle.fastdeploy.app.examples.R;
-import com.baidu.paddle.fastdeploy.app.ui.view.AppCompatPreferenceActivity;
+import com.baidu.paddle.fastdeploy.app.ui.AppCompatPreferenceActivity;
 import com.baidu.paddle.fastdeploy.app.ui.Utils;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class DetectionSettingsActivity extends AppCompatPreferenceActivity implements
+public class SettingsActivity extends AppCompatPreferenceActivity implements
         SharedPreferences.OnSharedPreferenceChangeListener {
-    private static final String TAG = DetectionSettingsActivity.class.getSimpleName();
+    private static final String TAG = SettingsActivity.class.getSimpleName();
 
     static public int selectedModelIdx = -1;
     static public String modelDir = "";
@@ -46,7 +46,7 @@ public class DetectionSettingsActivity extends AppCompatPreferenceActivity imple
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.detection_settings);
+        addPreferencesFromResource(R.xml.ocr_settings);
         ActionBar supportActionBar = getSupportActionBar();
         if (supportActionBar != null) {
             supportActionBar.setDisplayHomeAsUpEnabled(true);
@@ -59,8 +59,8 @@ public class DetectionSettingsActivity extends AppCompatPreferenceActivity imple
         preInstalledCPUPowerModes = new ArrayList<String>();
         preInstalledScoreThresholds = new ArrayList<String>();
         preInstalledEnableLiteFp16s = new ArrayList<String>();
-        preInstalledModelDirs.add(getString(R.string.MODEL_DIR_DEFAULT));
-        preInstalledLabelPaths.add(getString(R.string.LABEL_PATH_DEFAULT));
+        preInstalledModelDirs.add(getString(R.string.OCR_MODEL_DIR_DEFAULT));
+        preInstalledLabelPaths.add(getString(R.string.OCR_REC_LABEL_DEFAULT));
         preInstalledCPUThreadNums.add(getString(R.string.CPU_THREAD_NUM_DEFAULT));
         preInstalledCPUPowerModes.add(getString(R.string.CPU_POWER_MODE_DEFAULT));
         preInstalledScoreThresholds.add(getString(R.string.SCORE_THRESHOLD_DEFAULT));
@@ -90,7 +90,7 @@ public class DetectionSettingsActivity extends AppCompatPreferenceActivity imple
         SharedPreferences sharedPreferences = getPreferenceScreen().getSharedPreferences();
 
         String selected_model_dir = sharedPreferences.getString(getString(R.string.CHOOSE_PRE_INSTALLED_MODEL_KEY),
-                getString(R.string.MODEL_DIR_DEFAULT));
+                getString(R.string.OCR_MODEL_DIR_DEFAULT));
         int selected_model_idx = lpChoosePreInstalledModel.findIndexOfValue(selected_model_dir);
         if (selected_model_idx >= 0 && selected_model_idx < preInstalledModelDirs.size() && selected_model_idx != selectedModelIdx) {
             SharedPreferences.Editor editor = sharedPreferences.edit();
@@ -106,9 +106,9 @@ public class DetectionSettingsActivity extends AppCompatPreferenceActivity imple
         }
 
         String model_dir = sharedPreferences.getString(getString(R.string.MODEL_DIR_KEY),
-                getString(R.string.MODEL_DIR_DEFAULT));
+                getString(R.string.OCR_MODEL_DIR_DEFAULT));
         String label_path = sharedPreferences.getString(getString(R.string.LABEL_PATH_KEY),
-                getString(R.string.LABEL_PATH_DEFAULT));
+                getString(R.string.OCR_REC_LABEL_DEFAULT));
         String cpu_thread_num = sharedPreferences.getString(getString(R.string.CPU_THREAD_NUM_KEY),
                 getString(R.string.CPU_THREAD_NUM_DEFAULT));
         String cpu_power_mode = sharedPreferences.getString(getString(R.string.CPU_POWER_MODE_KEY),
@@ -136,12 +136,12 @@ public class DetectionSettingsActivity extends AppCompatPreferenceActivity imple
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(ctx);
 
         String model_dir = sharedPreferences.getString(ctx.getString(R.string.MODEL_DIR_KEY),
-                ctx.getString(R.string.MODEL_DIR_DEFAULT));
+                ctx.getString(R.string.OCR_MODEL_DIR_DEFAULT));
         settingsChanged |= !modelDir.equalsIgnoreCase(model_dir);
         modelDir = model_dir;
 
         String label_path = sharedPreferences.getString(ctx.getString(R.string.LABEL_PATH_KEY),
-                ctx.getString(R.string.LABEL_PATH_DEFAULT));
+                ctx.getString(R.string.OCR_REC_LABEL_DEFAULT));
         settingsChanged |= !labelPath.equalsIgnoreCase(label_path);
         labelPath = label_path;
