+++ java/android/app/src/main/java/com/baidu/paddle/fastdeploy/app/ui/ActionBarLayout.java
@@ -1,4 +1,4 @@
-package com.baidu.paddle.fastdeploy.app.ui.layout;
+package com.baidu.paddle.fastdeploy.app.ui;
 
 import android.content.Context;
 import android.graphics.Color;
