+++ java/android/app/src/main/java/com/baidu/paddle/fastdeploy/app/ui/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package com.baidu.paddle.fastdeploy.app.ui.view;
+package com.baidu.paddle.fastdeploy.app.ui;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
