+++ java/android/app/src/main/java/com/baidu/paddle/fastdeploy/app/ui/CameraSurfaceView.java
@@ -1,4 +1,4 @@
-package com.baidu.paddle.fastdeploy.app.ui.view;
+package com.baidu.paddle.fastdeploy.app.ui;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -14,8 +14,6 @@ import android.opengl.GLUtils;
 import android.opengl.Matrix;
 import android.util.AttributeSet;
 import android.util.Log;
-
-import com.baidu.paddle.fastdeploy.app.ui.Utils;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
