+++ java/android/app/src/main/res/layout/activity_main.xml
@@ -5,15 +5,15 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:keepScreenOn="true"
-    tools:context=".ocr.OcrMainActivity">
+    tools:context="com.baidu.paddle.fastdeploy.app.examples.ocr.MainActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/colorWindow">
 
-        <com.baidu.paddle.fastdeploy.app.ui.layout.ActionBarLayout
-            android:id="@+id/action_bar_main"
+        <com.baidu.paddle.fastdeploy.app.ui.ActionBarLayout
+            android:id="@+id/action_bar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
@@ -42,7 +42,7 @@
                     android:layout_height="wrap_content"
                     android:text="@string/action_bar_take_photo"
                     android:textAlignment="center"
-                    android:visibility="gone"/>
+                    android:visibility="gone" />
 
                 <TextView
                     android:id="@+id/action_realtime_btn"
@@ -52,15 +52,16 @@
                     android:text="@string/action_bar_realtime"
                     android:textAlignment="center" />
             </LinearLayout>
-        </com.baidu.paddle.fastdeploy.app.ui.layout.ActionBarLayout>
 
-        <!--        实时-->
-        <com.baidu.paddle.fastdeploy.app.ui.view.CameraSurfaceView
+        </com.baidu.paddle.fastdeploy.app.ui.ActionBarLayout>
+
+        <!--        时实-->
+        <com.baidu.paddle.fastdeploy.app.ui.CameraSurfaceView
             android:id="@+id/sv_preview"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_above="@+id/contral"
-            android:layout_below="@+id/action_bar_main"
+            android:layout_below="@+id/action_bar"
             android:layout_centerInParent="true" />
 
         <ImageView
@@ -72,7 +73,8 @@
             android:layout_marginRight="20dp"
             android:layout_marginBottom="145dp"
             android:background="@drawable/album_btn"
-            android:scaleType="fitXY" />
+            android:scaleType="fitXY"
+            android:visibility="gone"/>
 
         <TextView
             android:id="@+id/tv_status"
