--- java/android/app/src/main/res/layout/detection_activity_main.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <include
-        layout="@layout/detection_camera_page"
-        android:id="@+id/camera_page"></include>
-
-    <include
-        layout="@layout/detection_result_page"
-        android:id="@+id/result_page"
-        android:visibility="gone"></include>
-</FrameLayout>
