--- java/android/app/src/main/res/layout/detection_result_page_item.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@drawable/result_page_border_section_bk">
-
-    <TextView
-        android:id="@+id/index"
-        style="@style/list_result_view_item_style"
-        android:layout_width="wrap_content"
-        android:layout_weight="0.2" />
-
-    <TextView
-        android:id="@+id/name"
-        style="@style/list_result_view_item_style"
-        android:layout_width="wrap_content"
-        android:layout_weight="0.6"
-        android:maxWidth="300px" />
-
-    <TextView
-        android:id="@+id/confidence"
-        style="@style/list_result_view_item_style"
-        android:layout_weight="0.2"
-        android:layout_width="wrap_content" />
-</LinearLayout>
