+++ java/android/app/src/main/res/values/strings.xml
@@ -1,9 +1,9 @@
 <resources>
     <!-- Default App name -->
-    <string name="app_name">EasyEdge</string>
+    <string name="app_name">FastDeploy PicoDet</string>
     <!-- Other App name -->
-    <string name="detection_app_name">EasyEdge</string>
-    <string name="ocr_app_name">EasyEdge</string>
+    <string name="detection_app_name">FastDeploy PicoDet</string>
+    <string name="ocr_app_name">FastDeploy PP-OCRv2</string>
     <!-- Keys for PreferenceScreen -->
     <string name="CHOOSE_PRE_INSTALLED_MODEL_KEY">CHOOSE_INSTALLED_MODEL_KEY</string>
     <string name="MODEL_DIR_KEY">MODEL_DIR_KEY</string>
@@ -17,7 +17,6 @@
     <string name="CPU_POWER_MODE_DEFAULT">LITE_POWER_HIGH</string>
     <string name="SCORE_THRESHOLD_DEFAULT">0.4</string>
     <string name="ENABLE_LITE_FP16_MODE_DEFAULT">true</string>
-    <!--Other values-->
     <!-- Default model & Label paths & other values ... -->
     <string name="MODEL_DIR_DEFAULT">models/picodet_s_320_coco_lcnet</string>
     <string name="LABEL_PATH_DEFAULT">labels/coco_label_list.txt</string>
