+++ java/android/app/src/main/res/xml/settings.xml
